up data on news type activity stream with no target

+++ controllers/DataMigrationController.php
@@ -180,14 +180,26 @@ class DataMigrationController extends CommunecterController {
    	
    	public function actionWashingNewsNoTarget(){
   		$news=PHDB::find(News::COLLECTION);
+  		$i=0;
+  		$nbNews=count($news);
+		echo "Nombre de documents appelés : ".$nbNews;
   		foreach($news as $key => $data){
-		  if(!@$data["target"]){
+		  if($data["type"]=="news" && !@$data["target"]){
+			  $i++;
+			  print_r($data);
+			  PHDB::remove(News::COLLECTION, array("_id"=>new MongoId($key)));
+		 // PHDB::remove(News::COLLECTION, array("_id"=>new MongoId($key)));
+		  	
+			}
+			else if ($data["type"]=="activityStream" && !@$data["target"]){
+			$i++;
 			  print_r($data);
 			  PHDB::remove(News::COLLECTION, array("_id"=>new MongoId($key)));
 		 // PHDB::remove(News::COLLECTION, array("_id"=>new MongoId($key)));
 		  	
 			}
 		}
+		echo "nombre de news traitées ////////////// ".$i;
 	}
 	// Delete news with object gantts and needs
 	public function actionDeleteNewsGanttsNeeds(){
